pyproject.toml
[tool.poetry]
name = "avatarme"
version = "0.1.0"
description = ""
authors = ["Badr Ben m'barek <badr.benb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pillow = "^10.1.0"
replicate = "^0.20.0"
google-cloud-storage = "^2.13.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
ipykernel = "^6.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
